refactor: deduplicate git root resolution into edda-core (#236)
Move resolve_git_root from edda-store and edda-ledger into edda-core::git to eliminate identical duplicate implementations. Both crates now call the shared function, preventing future divergence that could cause silent data loss via mismatched project IDs.
