+++ crates/edda-store/Cargo.toml
@@ -10,6 +10,7 @@ categories.workspace = true
 keywords.workspace = true
 
 [dependencies]
+edda-core = { path = "../edda-core", version = "0.1.1" }
 anyhow.workspace = true
 thiserror.workspace = true
 serde.workspace = true
